requirements.txt
fastapi
openai>=1.1.1
python-dotenv
uvicorn